+++ lib/mongoid/shell/commands.rb
@@ -1,2 +1,3 @@
 require 'mongoid/shell/commands/base'
+require 'mongoid/shell/commands/mongo'
 require 'mongoid/shell/commands/mongodump'