+++ lib/mongoid/shell/errors.rb
@@ -1,3 +1,4 @@
 require 'mongoid/shell/errors/base'
 require 'mongoid/shell/errors/missing_session_error'
 require 'mongoid/shell/errors/session_not_connected_error'
+require 'mongoid/shell/errors/missing_primary_node_error'