+++ lib/mongoid/shell/errors/missing_primary_node_error.rb
@@ -1,0 +1,11 @@
+module Mongoid
+  module Shell
+    module Errors
+      class MissingPrimaryNodeError < Mongoid::Shell::Errors::Base
+        def initialize
+          super(compose_message("missing_primary_node"))
+        end
+      end
+    end
+  end
+end