+++ lib/mongoid/shell/properties.rb
@@ -1,4 +1,5 @@
 require 'mongoid/shell/properties/host'
+require 'mongoid/shell/properties/primary'
 require 'mongoid/shell/properties/database'
 require 'mongoid/shell/properties/username'
 require 'mongoid/shell/properties/password'